feat: Changed offer endpoints.
