+++ src/handlers/sra_handlers.ts
@@ -145,84 +145,6 @@ export class SRAHandlers {
         }
         ORDERS_POST_REQUESTS.labels('multi', CHAIN_ID.toString()).inc();
     }
-    public async offerCreateContingentPoolsAsync(req: express.Request, res: express.Response): Promise<void> {
-        const params = offerCreateContingentPoolFilterParams(req);
-
-        const response = await this._orderBook.offerCreateContingentPoolsAsync(params);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async getOfferCreateContingentPoolByOfferHashAsync(
-        req: express.Request,
-        res: express.Response,
-    ): Promise<void> {
-        const response = await this._orderBook.getOfferCreateContingentPoolByOfferHashAsync(req.params.offerHash);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async postOfferCreateContingentPoolAsync(req: express.Request, res: express.Response): Promise<void> {
-        schemaUtils.validateSchema(req.body, schemas.sraOfferCreateContingentPoolSchema);
-
-        const offerCreateContingentPoolEntity = new OfferCreateContingentPoolEntity(req.body);
-        const response = await this._orderBook.postOfferCreateContingentPoolAsync(offerCreateContingentPoolEntity);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async offerAddLiquidityAsync(req: express.Request, res: express.Response): Promise<void> {
-        const params = offerLiquidityFilterParams(req);
-        const response = await this._orderBook.offerAddLiquidityAsync(params);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async getOfferAddLiquidityByOfferHashAsync(req: express.Request, res: express.Response): Promise<void> {
-        const response = await this._orderBook.getOfferAddLiquidityByOfferHashAsync(req.params.offerHash);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async postOfferAddLiquidityAsync(req: express.Request, res: express.Response): Promise<void> {
-        schemaUtils.validateSchema(req.body, schemas.sraOfferAddLiquiditySchema);
-
-        const offerAddLiquidityEntity = new OfferAddLiquidityEntity({
-            ...req.body,
-            referenceAsset: NULL_TEXT,
-            collateralToken: NULL_ADDRESS,
-            dataProvider: NULL_ADDRESS,
-            permissionedERC721Token: NULL_ADDRESS,
-        });
-
-        const response = await this._orderBook.postOfferLiquidityAsync(offerAddLiquidityEntity, OfferLiquidityType.Add);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async offerRemoveLiquidityAsync(req: express.Request, res: express.Response): Promise<void> {
-        const params = offerLiquidityFilterParams(req);
-        const response = await this._orderBook.offerRemoveLiquidityAsync(params);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async getOfferRemoveLiquidityByOfferHashAsync(req: express.Request, res: express.Response): Promise<void> {
-        const response = await this._orderBook.getOfferRemoveLiquidityByOfferHashAsync(req.params.offerHash);
-
-        res.status(HttpStatus.OK).send(response);
-    }
-    public async postOfferRemoveLiquidityAsync(req: express.Request, res: express.Response): Promise<void> {
-        schemaUtils.validateSchema(req.body, schemas.sraOfferRemoveLiquiditySchema);
-
-        const offerRemoveLiquidityEntity = new OfferRemoveLiquidityEntity({
-            ...req.body,
-            referenceAsset: NULL_TEXT,
-            collateralToken: NULL_ADDRESS,
-            dataProvider: NULL_ADDRESS,
-            permissionedERC721Token: NULL_ADDRESS,
-        });
-
-        const response = await this._orderBook.postOfferLiquidityAsync(
-            offerRemoveLiquidityEntity,
-            OfferLiquidityType.Remove,
-        );
-
-        res.status(HttpStatus.OK).send(response);
-    }
     public async postPersistentOrderAsync(req: express.Request, res: express.Response): Promise<void> {
         const shouldSkipConfirmation = req.query.skipConfirmation === 'true';
         const apiKey = req.header('0x-api-key');
@@ -244,45 +166,6 @@ export class SRAHandlers {
             res.status(StatusCodes.OK).send();
         }
     }
-}
-
-// The function to get filter parameter about OfferCreateContingentPool
-function offerCreateContingentPoolFilterParams(req: express.Request): any {
-    const { page, perPage } = paginationUtils.parsePaginationConfig(req);
-    const maker = req.query.maker === undefined ? NULL_ADDRESS : (req.query.maker as string).toLowerCase();
-    const taker = req.query.taker === undefined ? NULL_ADDRESS : (req.query.taker as string).toLowerCase();
-    const makerDirection = req.query.makerDirection === undefined ? NULL_TEXT : (req.query.makerDirection as string);
-    const referenceAsset = req.query.referenceAsset === undefined ? NULL_TEXT : (req.query.referenceAsset as string);
-    const collateralToken =
-        req.query.collateralToken === undefined ? NULL_ADDRESS : (req.query.collateralToken as string).toLowerCase();
-    const dataProvider =
-        req.query.dataProvider === undefined ? NULL_ADDRESS : (req.query.dataProvider as string).toLowerCase();
-    const permissionedERC721Token =
-        req.query.permissionedERC721Token === undefined
-            ? NULL_ADDRESS
-            : (req.query.permissionedERC721Token as string).toLowerCase();
-
-    return {
-        page,
-        perPage,
-        maker,
-        taker,
-        makerDirection,
-        referenceAsset,
-        collateralToken,
-        dataProvider,
-        permissionedERC721Token,
-    };
-}
-
-// The function to get filter parameter about OfferAddLiqudity or OfferRemoveLiqudity
-function offerLiquidityFilterParams(req: express.Request): any {
-    const params = offerCreateContingentPoolFilterParams(req);
-
-    return {
-        ...params,
-        poolId: req.query.poolId === undefined ? NULL_TEXT : (req.query.poolId as string),
-    };
 }
 
 function validateAssetTokenOrThrow(allowedTokens: string[], tokenAddress: string, field: string): void {
